+++ CHANGES.md
@@ -12,6 +12,7 @@ Changes in 8.1
 - SQL ACCOUNTING_SALARIES table: Add FILE_ATTACHED column in Update.fnc.php & rosariosis.sql
 - Add File Attached to Salaries in Salaries.php & Accounting/functions.inc.php
 - SQL BILLING_FEES table: Add FILE_ATTACHED column in Update.fnc.php & rosariosis.sql
+- Add File Attached to Fees in StudentFees.php & Student_Billing/functions.inc.php
 
 Changes in 8.0.4
 ----------------